requirements.txt
amazon-braket-sdk
tensorflow
h5py==2.10.0